+++ server/src/main/kotlin/com/adobe/testing/s3mock/controller/BucketController.kt
@@ -1,5 +1,5 @@
 /*
- *  Copyright 2017-2025 Adobe.
+ *  Copyright 2017-2026 Adobe.
  *
  *  Licensed under the Apache License, Version 2.0 (the "License");
  *  you may not use this file except in compliance with the License.
@@ -26,6 +26,7 @@ import com.adobe.testing.s3mock.dto.LocationConstraint
 import com.adobe.testing.s3mock.dto.ObjectLockConfiguration
 import com.adobe.testing.s3mock.dto.ObjectOwnership
 import com.adobe.testing.s3mock.dto.Region
+import com.adobe.testing.s3mock.dto.Tagging
 import com.adobe.testing.s3mock.dto.VersioningConfiguration
 import com.adobe.testing.s3mock.service.BucketService
 import com.adobe.testing.s3mock.util.AwsHttpHeaders.X_AMZ_BUCKET_OBJECT_LOCK_ENABLED
@@ -45,11 +46,13 @@ import com.adobe.testing.s3mock.util.AwsHttpParameters.NOT_LIFECYCLE
 import com.adobe.testing.s3mock.util.AwsHttpParameters.NOT_LIST_TYPE
 import com.adobe.testing.s3mock.util.AwsHttpParameters.NOT_LOCATION
 import com.adobe.testing.s3mock.util.AwsHttpParameters.NOT_OBJECT_LOCK
+import com.adobe.testing.s3mock.util.AwsHttpParameters.NOT_TAGGING
 import com.adobe.testing.s3mock.util.AwsHttpParameters.NOT_UPLOADS
 import com.adobe.testing.s3mock.util.AwsHttpParameters.NOT_VERSIONING
 import com.adobe.testing.s3mock.util.AwsHttpParameters.NOT_VERSIONS
 import com.adobe.testing.s3mock.util.AwsHttpParameters.OBJECT_LOCK
 import com.adobe.testing.s3mock.util.AwsHttpParameters.START_AFTER
+import com.adobe.testing.s3mock.util.AwsHttpParameters.TAGGING
 import com.adobe.testing.s3mock.util.AwsHttpParameters.VERSIONING
 import com.adobe.testing.s3mock.util.AwsHttpParameters.VERSIONS
 import com.adobe.testing.s3mock.util.AwsHttpParameters.VERSION_ID_MARKER
@@ -119,6 +122,7 @@ class BucketController(private val bucketService: BucketService) {
     params = [
       NOT_OBJECT_LOCK,
       NOT_LIFECYCLE,
+      NOT_TAGGING,
       NOT_VERSIONING
     ]
   )
@@ -186,7 +190,8 @@ class BucketController(private val bucketService: BucketService) {
       // AWS SDK V1 pattern
       "/{bucketName:.+}/"
     ], params = [
-      NOT_LIFECYCLE
+      NOT_LIFECYCLE,
+      NOT_TAGGING
     ]
   )
   @S3Verified(year = 2025)
@@ -366,6 +371,76 @@ class BucketController(private val bucketService: BucketService) {
   }
 
   /**
+   * [API Reference](https://docs.aws.amazon.com/AmazonS3/latest/API/API_GetBucketTagging.html).
+   */
+  @GetMapping(
+    value = [
+      // AWS SDK V2 pattern
+      "/{bucketName:.+}",
+      // AWS SDK V1 pattern
+      "/{bucketName:.+}/"
+    ],
+    params = [
+      TAGGING,
+      NOT_LIST_TYPE
+    ],
+    produces = [
+      MediaType.APPLICATION_XML_VALUE
+    ]
+  )
+  @S3Verified(year = 2026)
+  fun getBucketTagging(@PathVariable bucketName: String): ResponseEntity<Tagging> {
+    bucketService.verifyBucketExists(bucketName)
+    val tagging = bucketService.getBucketTagging(bucketName)
+    return ResponseEntity.ok(tagging)
+  }
+
+  /**
+   * [API Reference](https://docs.aws.amazon.com/AmazonS3/latest/API/API_PutBucketTagging.html).
+   */
+  @PutMapping(
+    value = [
+      // AWS SDK V2 pattern
+      "/{bucketName:.+}",
+      // AWS SDK V1 pattern
+      "/{bucketName:.+}/"
+    ],
+    params = [
+      TAGGING
+    ]
+  )
+  @S3Verified(year = 2026)
+  fun putBucketTagging(
+    @PathVariable bucketName: String,
+    @RequestBody tagging: Tagging
+  ): ResponseEntity<Void> {
+    bucketService.verifyBucketExists(bucketName)
+    bucketService.setBucketTagging(bucketName, tagging)
+    return ResponseEntity.ok().build()
+  }
+
+  /**
+   * [API Reference](https://docs.aws.amazon.com/AmazonS3/latest/API/API_DeleteBucketTagging.html).
+   */
+  @DeleteMapping(
+    value = [
+      // AWS SDK V2 pattern
+      "/{bucketName:.+}",
+      // AWS SDK V1 pattern
+      "/{bucketName:.+}/"
+    ],
+    params = [
+      TAGGING
+    ]
+  )
+  @S3Verified(year = 2026)
+  fun deleteBucketTagging(@PathVariable bucketName: String): ResponseEntity<Void> {
+    bucketService.verifyBucketExists(bucketName)
+    bucketService.deleteBucketTagging(bucketName)
+    return ResponseEntity.noContent().build()
+  }
+
+  /**
    * [API Reference](https://docs.aws.amazon.com/AmazonS3/latest/API/API_GetBucketLocation.html).
    */
   @GetMapping(
@@ -404,6 +479,7 @@ class BucketController(private val bucketService: BucketService) {
       NOT_OBJECT_LOCK,
       NOT_LIST_TYPE,
       NOT_LIFECYCLE,
+      NOT_TAGGING,
       NOT_LOCATION,
       NOT_VERSIONS,
       NOT_VERSIONING